PFW-880 Allow menu edit items to use 1B variables
Change in memory:
Flash: -526 bytes
SRAM: -5 bytes

+++ Firmware/backlight.h
@@ -13,8 +13,8 @@ enum Backlight_Mode
 	BACKLIGHT_MODE_AUTO    = 2,
 };
 
-extern int16_t backlightLevel_HIGH;
-extern int16_t backlightLevel_LOW;
+extern uint8_t backlightLevel_HIGH;
+extern uint8_t backlightLevel_LOW;
 extern uint8_t backlightMode;
 extern bool backlightSupport;
 extern int16_t backlightTimer_period;
